MoreFrontMethods

+++ src/main/java/tn/esprit/esprithub/services/SportTeamService.java
@@ -2,9 +2,12 @@ package tn.esprit.esprithub.services;
 
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.EntityNotFoundException;
+import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
+import org.springframework.web.multipart.MultipartFile;
 import tn.esprit.esprithub.entities.Field;
 import tn.esprit.esprithub.entities.Reservation;
 import tn.esprit.esprithub.entities.SportTeam;
@@ -14,6 +17,11 @@ import tn.esprit.esprithub.repository.IReservationRepository;
 import tn.esprit.esprithub.repository.ISportTeamRepository;
 import tn.esprit.esprithub.repository.IUserRepository;
 
+import java.io.IOException;
+import java.io.OutputStream;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.*;
 
 @Service
@@ -39,7 +47,18 @@ public class SportTeamService implements ISportTeamService{
 
     @Override
     public void deleteSportTeam(Long sportTeamId) {
-        sportTeamRepository.deleteById(sportTeamId);
+        SportTeam sportTeam = sportTeamRepository.findById(sportTeamId).orElse(null);
+
+        if (sportTeam != null) {
+            for (User user : sportTeam.getUsers()) {
+                user.setSportTeams(null);
+            }
+            sportTeam.setCaptain(null);
+            sportTeam.getUsers().clear();
+
+
+            sportTeamRepository.deleteById(sportTeamId);
+        }
     }
 
     @Override
@@ -54,7 +73,139 @@ public class SportTeamService implements ISportTeamService{
 
 
 
+    @Override
+    public SportTeam addSportTeamCap2(SportTeam sportTeam, Long captainId, MultipartFile photoFile) {
+        User user = userRepository.findById(captainId).orElse(null);
+        String uploadPath = "C:\\Users\\Bacem\\IdeaProjects\\EspritHub\\src\\main\\resources\\static\\photos\\";
+        try {
+            byte[] bytes = photoFile.getBytes();
+            String fileName = photoFile.getOriginalFilename();
+            Path path = Paths.get(uploadPath + fileName);
+            Files.write(path, bytes);
+            sportTeam.setLogoTeam(fileName);
+            sportTeam.setCaptain(user);
 
+            user.setSportTeams(sportTeam);
+            sportTeamRepository.save(sportTeam);
+
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+
+
+        return sportTeamRepository.save(sportTeam);
+    }
+//@Override
+//public SportTeam addSportTeamCap3(String teamName, Long captainId, MultipartFile photoFile) {
+//    SportTeam sportTeam = new SportTeam();
+//
+//    User user = userRepository.findById(captainId).orElse(null);
+//
+//    String uploadPath = "C:\\Users\\Bacem\\IdeaProjects\\EspritHub\\src\\main\\resources\\static\\photos\\";
+//    try {
+//        byte[] bytes = photoFile.getBytes();
+//        String fileName = photoFile.getOriginalFilename();
+//        Path path = Paths.get(uploadPath + fileName);
+//        Files.write(path, bytes);
+//
+//        sportTeam.setNameTeam(teamName);
+//        sportTeam.setLogoTeam(fileName);
+//        sportTeam.setCaptain(user);
+//
+//        user.setSportTeams(sportTeam);
+//        sportTeamRepository.save(sportTeam);
+//    } catch (IOException e) {
+//        e.printStackTrace();
+//    }
+//    return sportTeam;
+//}
+@Override
+@Transactional
+public SportTeam addSportTeamCap3(String teamName, Long captainId, MultipartFile photoFile) {
+    SportTeam sportTeam = new SportTeam();
+
+    User user = userRepository.findById(captainId).orElseThrow(() -> new EntityNotFoundException("User not found"));
+
+    String uploadPath = "C:\\Users\\Bacem\\IdeaProjects\\EspritHub\\src\\main\\resources\\static\\photos\\";
+    try {
+        byte[] bytes = photoFile.getBytes();
+        String fileName = StringUtils.cleanPath(photoFile.getOriginalFilename());
+        Path path = Paths.get(uploadPath, fileName);
+
+        Files.createDirectories(path.getParent());
+
+        try (OutputStream os = Files.newOutputStream(path)) {
+            os.write(bytes);
+        }
+        sportTeam.setNameTeam(teamName);
+        sportTeam.setLogoTeam(fileName);
+        sportTeam.setCaptain(user);
+        user.setSportTeams(sportTeam);
+        userRepository.save(user);
+        sportTeamRepository.save(sportTeam);
+
+    } catch (IOException e) {
+
+
+    }
+    return sportTeam;
+}
+
+    @Override
+    public SportTeam updateSportTeamCapWithPhoto(String teamName, Long sportTeamId, MultipartFile photoFile) {
+        SportTeam existingSportTeam = sportTeamRepository.findById(sportTeamId).orElse(null);
+        if (existingSportTeam != null) {
+            existingSportTeam.setNameTeam(teamName);
+            if (photoFile != null && !photoFile.isEmpty()) {
+                try {
+                    String uploadPath = "C:\\Users\\Bacem\\IdeaProjects\\EspritHub\\src\\main\\resources\\static\\photos\\";
+                    byte[] bytes = photoFile.getBytes();
+                    String fileName = StringUtils.cleanPath(photoFile.getOriginalFilename());
+                    Path path = Paths.get(uploadPath, fileName);
+                    Files.createDirectories(path.getParent());
+                    try (OutputStream os = Files.newOutputStream(path)) {
+                        os.write(bytes);
+                    }
+                    existingSportTeam.setLogoTeam(fileName);
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+            return sportTeamRepository.save(existingSportTeam);
+        }
+        return null;
+    }
+
+
+
+//    @Override
+//    public SportTeam updateSportTeamCapWithPhoto(String teamName, Long sportTeamId, MultipartFile photoFile) {
+//        SportTeam existingSportTeam = sportTeamRepository.findById(sportTeamId).orElse(null);
+//        if (existingSportTeam != null) {
+//            existingSportTeam.setNameTeam(teamName);
+//
+//            if (photoFile != null && !photoFile.isEmpty()) {
+//                try {
+//                    String uploadPath = "C:\\Users\\Bacem\\IdeaProjects\\EspritHub\\src\\main\\resources\\static\\photos\\";
+//
+//                    byte[] bytes = photoFile.getBytes();
+//
+//                    String fileName = photoFile.getOriginalFilename();
+//
+//                    Path path = Paths.get(uploadPath + fileName);
+//
+//                    Files.write(path, bytes);
+//                    existingSportTeam.setLogoTeam(fileName);
+//
+//                } catch (IOException e) {
+//                    e.printStackTrace();
+//                }
+//            }
+//
+//            return sportTeamRepository.save(existingSportTeam);
+//        }
+//        return null;
+//    }
 
 
     @Override
@@ -79,22 +230,6 @@ public class SportTeamService implements ISportTeamService{
         return sportTeamRepository.save(existingSportTeam);
     }
 
-//    @Override
-//    public void deleteSportTeamCap(Long sportTeamId) {
-//        SportTeam sportTeam = sportTeamRepository.findById(sportTeamId).orElse(null);
-//
-//        if (sportTeam != null) {
-//            Long captainId = 5L;
-//
-//            if (sportTeam.getCaptain().getUserId().equals(captainId)) {
-//                sportTeamRepository.deleteById(sportTeamId);
-//            } else {
-//                throw new EntityNotFoundException("Only the captain can delete the sport team.");
-//            }
-//        } else {
-//            throw new EntityNotFoundException("Sport team not found with ID: " + sportTeamId);
-//        }
-//    }
 
     @Override
     public void deleteSportTeamCap(Long sportTeamId, Long captainId) {
@@ -102,7 +237,6 @@ public class SportTeamService implements ISportTeamService{
 
         if (sportTeam != null) {
             if (sportTeam.getCaptain().getUserId().equals(captainId)) {
-                // Dissociate users from the sport team
                 for (User user : sportTeam.getUsers()) {
                     user.setSportTeams(null);
                 }
@@ -139,6 +273,33 @@ public class SportTeamService implements ISportTeamService{
 
         }
     }
+
+    @Override
+    public void addUserByEmailToSportTeam(Long sportTeamId, String userEmail) {
+        User user = userRepository.findByEmail(userEmail);
+        SportTeam sportTeam = sportTeamRepository.findById(sportTeamId).orElse(null);
+
+        if (user != null && sportTeam != null) {
+            sportTeam.getUsers().add(user);
+            user.setSportTeams(sportTeam);
+            sportTeamRepository.save(sportTeam);
+        }
+    }
+
+    @Override
+    public void RemoveUserByEmailFromSportTeamE(Long sportTeamId, String userEmail) {
+        User user = userRepository.findByEmail(userEmail);
+        SportTeam sportTeam = sportTeamRepository.findById(sportTeamId).orElse(null);
+
+        if (user != null && sportTeam != null) {
+            sportTeam.getUsers().remove(user);
+            user.setSportTeams(null); // Remove the reference to the sport team from the user
+            userRepository.save(user);
+            sportTeamRepository.save(sportTeam);
+
+        }
+    }
+
 
 
     @Override
@@ -190,54 +351,7 @@ public class SportTeamService implements ISportTeamService{
     public List<User> getUsersBySportTeamId(Long sportTeamId) {
         return sportTeamRepository.findUserIdsBySportTeamId(sportTeamId);
     }
-//    @Override
-//    public void makeTeamReservation(Long sportTeamId, Long captainId, Long fieldId, Reservation reservation) {
-//        SportTeam sportTeam = sportTeamRepository.findById(sportTeamId).orElse(null);
-//        Field field = fieldRepository.findById(fieldId).orElse(null);
-//
-//        if (sportTeam != null && sportTeam.getCaptain().getUserId().equals(captainId) && field != null) {
-//            Set<User> teamMembers = sportTeam.getUsers();
-//
-//            for (User member : teamMembers) {
-//                reservation.setUsers(new HashSet<>(Collections.singletonList(member)));
-//                reservation.setFields(field);
-//                member.getReservations().add(reservation);
-//            }
-//
-//            userRepository.saveAll(teamMembers);
-//        } else {
-//            throw new IllegalArgumentException("Only the captain can make a team reservation and the provided field must exist.");
-//        }
-//    }
 
-
-//    @Override
-//    public void makeTeamReservation(Long sportTeamId, Long captainId, Long fieldId, Reservation reservation) {
-//        SportTeam sportTeam = sportTeamRepository.findById(sportTeamId).orElse(null);
-//        Field field = fieldRepository.findById(fieldId).orElse(null);
-//
-//        if (sportTeam != null && sportTeam.getCaptain().getUserId().equals(captainId) && field != null) {
-//            Set<User> teamMembers = sportTeam.getUsers();
-//
-//            Reservation teamReservation = new Reservation();
-//            teamReservation.setStartDate(reservation.getStartDate());
-//            teamReservation.setEndDate(reservation.getEndDate());
-//            teamReservation.setNbPlayers(reservation.getNbPlayers());
-//            teamReservation.setResStatus(reservation.getResStatus());
-//            teamReservation.setResType(reservation.getResType());
-//            teamReservation.setFields(field);
-//
-//            for (User member : teamMembers) {
-//                member.getReservations().add(teamReservation);
-//                reservationRepository.save(teamReservation);
-//            }
-//
-//            userRepository.saveAll(teamMembers);
-//        } else {
-//            throw new IllegalArgumentException("Only the captain can make a team reservation and the provided field must exist.");
-//        }
-//    }
-//
 @Override
 public void makeTeamReservation(Long sportTeamId, Long captainId, Long fieldId, Reservation reservation) {
     SportTeam sportTeam = sportTeamRepository.findById(sportTeamId).orElse(null);
@@ -258,6 +372,9 @@ public void makeTeamReservation(Long sportTeamId, Long captainId, Long fieldId, 
             member.getReservations().add(teamReservation);
         }
 
+        long nbPlayers = teamMembers.size();
+        teamReservation.setNbPlayers(nbPlayers);
+
         reservationRepository.save(teamReservation);
         userRepository.saveAll(teamMembers);
     } else {
@@ -265,7 +382,37 @@ public void makeTeamReservation(Long sportTeamId, Long captainId, Long fieldId, 
     }
 }
 
+    @Override
+    public int countUsersJoinedInSportTeam(Long teamId) {
+        return sportTeamRepository.countUsersJoinedInSportTeam(teamId);
+    }
+    @Override
+    public boolean isUserCaptain(Long userId) {
+        // Query the database to check if the user is a captain in any sport team
+        List<SportTeam> sportTeams = sportTeamRepository.findByCaptainUserId(userId);
+        return !sportTeams.isEmpty();
+    }
+    @Override
+    public boolean isUserCaptainTeam(Long teamId, Long userId) {
+        SportTeam sportTeam = sportTeamRepository.findById(teamId).orElse(null);
+        User user = userRepository.findById(userId).orElse(null);
+        if (sportTeam != null && user != null) {
+            if (sportTeam.getCaptain().getUserId() ==userId) {
+                return true;
+            }
+        }
+        return false;
+    }
 
 
+    @Override
+    public Long getSportTeamIdByCaptainId(Long captainId) {
+        SportTeam sportTeam = sportTeamRepository.findSportTeamsByCaptainUserId(captainId);
 
+        if (sportTeam != null) {
+            return sportTeam.getTeamId();
+        } else {
+            return null; // or throw an exception based on your error handling strategy
+        }
+    }
 }
